Update side panel description to def m
